New version for MapComponent
Nominatim code separated from mapComponent to allow easier use of other geocoder .
New return errors from Nominatim added to i18n files
Change alert to translate accept buton
