+++ .gitignore
@@ -150,3 +150,4 @@ dist-ssr
 *.sln
 *.sw?
 src/Simulator/log/PVGIScalls.txt
+src/Simulator/log/PVGIScalls.txt
